Initial Population
The Environmental Data Retrieval standard is to be built on OGC API-Common.  Both API-Records and API-Coverages are also built on API-Common.  This commit reuses content from that previous work as a foundation for the EDR API standard. We hope that this will promote consistency in the way API standards build on other API standards.